requirements.txt
requests>=2.31.0
rank_bm25>=0.2.2
tqdm>=4.66.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
